+++ tasks/queue-granules/index.js
@@ -8,7 +8,10 @@ const cumulusMessageAdapter = require('@cumulus/cumulus-message-adapter-js');
 const { enqueueGranuleIngestMessage } = require('@cumulus/ingest/queue');
 const { constructCollectionId } = require('@cumulus/message/Collections');
 const { buildExecutionArn } = require('@cumulus/message/Executions');
-const { providers: providersApi } = require('@cumulus/api-client');
+const {
+  providers: providersApi,
+  granules: granulesApi,
+} = require('@cumulus/api-client');
 const CollectionConfigStore = require('@cumulus/collection-config-store');
 
 async function fetchGranuleProvider(prefix, providerId) {
@@ -44,12 +47,12 @@ function groupAndBatchGranules(granules, batchSize = 1) {
 exports.groupAndBatchGranules = groupAndBatchGranules;
 
 /**
-* See schemas/input.json and schemas/config.json for detailed event description
-*
-* @param {Object} event - Lambda event object
-* @returns {Promise} - see schemas/output.json for detailed output schema
-*   that is passed to the next task in the workflow
-**/
+ * See schemas/input.json and schemas/config.json for detailed event description
+ *
+ * @param {Object} event - Lambda event object
+ * @returns {Promise} - see schemas/output.json for detailed output schema
+ *   that is passed to the next task in the workflow
+ **/
 async function queueGranules(event) {
   const granules = event.input.granules || [];
 
@@ -59,7 +62,8 @@ async function queueGranules(event) {
   );
 
   const arn = buildExecutionArn(
-    get(event, 'cumulus_config.state_machine'), get(event, 'cumulus_config.execution_name')
+    get(event, 'cumulus_config.state_machine'),
+    get(event, 'cumulus_config.execution_name')
   );
 
   const groupedAndBatchedGranules = groupAndBatchGranules(
@@ -74,7 +78,7 @@ async function queueGranules(event) {
         granuleBatch[0].dataType,
         granuleBatch[0].version
       );
-      return enqueueGranuleIngestMessage({
+      const executionArn = enqueueGranuleIngestMessage({
         granules: granuleBatch,
         queueUrl: event.config.queueUrl,
         granuleIngestWorkflow: event.config.granuleIngestWorkflow,
@@ -89,6 +93,18 @@ async function queueGranules(event) {
         executionNamePrefix: event.config.executionNamePrefix,
         additionalCustomMeta: event.config.childWorkflowMeta,
       });
+      if (executionArn) {
+        await Promise.all(granuleBatch.map((queuedGranule) =>
+          granulesApi.updateGranule({
+            prefix: event.config.stackName,
+            body: {
+              granuleId: queuedGranule.granuleId,
+              status: 'queued',
+              retries: 3,
+            },
+          })));
+      }
+      return executionArn;
     },
     { concurrency: get(event, 'config.concurrency', 3) }
   );
@@ -108,6 +124,10 @@ exports.queueGranules = queueGranules;
  *                              See schemas/output.json for detailed output schema
  */
 async function handler(event, context) {
-  return await cumulusMessageAdapter.runCumulusTask(queueGranules, event, context);
+  return await cumulusMessageAdapter.runCumulusTask(
+    queueGranules,
+    event,
+    context
+  );
 }
 exports.handler = handler;